--- GraphicsEngine/Graphics/Console/IConsoleGraphicsFrame.cs
@@ -1,10 +1,0 @@
-﻿namespace GraphicsEngine.Graphics.Console
-{
-	public interface IConsoleGraphicsFrame
-	{
-		int Width { get; }
-		int Height { get; }
-		IConsoleCharacterBuffer CharacterBuffer { get; }
-		IConsoleColorBuffer ColorBuffer { get; }
-	}
-}
