+++ GraphicsEngine/Graphics/GraphicsFrame.cs
@@ -1,38 +1,38 @@
-﻿namespace GraphicsEngine.Graphics.Console
+﻿namespace GraphicsEngine.Graphics
 {
-	public class ConsoleGraphicsFrame
-		: IConsoleGraphicsFrame
+	public class GraphicsFrame
+		: IGraphicsFrame
 	{
 		public const byte DefaultClearCharacter = 32;
 		public const byte DefaultClearColor = 0;
 
-		public ConsoleGraphicsFrame(int width, int height)
+		public GraphicsFrame(int width, int height)
 		{
 			Width = width;
 			Height = height;
 
 			ZBuffer = new float?[width, height];
-			CharacterBuffer = new ConsoleCharacterBuffer(width, height);
-			ColorBuffer = new ConsoleColorBuffer(width, height);
+			CharacterBuffer = new CharacterBuffer(width, height);
+			ColorBuffer = new ColorBuffer(width, height);
 		}
 
 		public float?[,] ZBuffer { get; private set; }
-		public ConsoleCharacterBuffer CharacterBuffer { get; }
-		public ConsoleColorBuffer ColorBuffer { get; }
+		public CharacterBuffer CharacterBuffer { get; }
+		public ColorBuffer ColorBuffer { get; }
 		public int Width { get; }
 		public int Height { get; }
 
-		IConsoleCharacterBuffer IConsoleGraphicsFrame.CharacterBuffer
+		ICharacterBuffer IGraphicsFrame.CharacterBuffer
 		{
 			get { return CharacterBuffer; }
 		}
 
-		IConsoleColorBuffer IConsoleGraphicsFrame.ColorBuffer
+		IColorBuffer IGraphicsFrame.ColorBuffer
 		{
 			get { return ColorBuffer; }
 		}
 
-		public void ClearBuffers(byte? clearCharacter = null, byte? clearColor = null)
+		public void Reset(byte? clearCharacter = null, byte? clearColor = null)
 		{
 			for (var x = 0; x < Width; x++)
 			{
