+++ GraphicsEngine/Graphics/IGraphicsFrame.cs
@@ -1,0 +1,10 @@
+﻿namespace GraphicsEngine.Graphics
+{
+	public interface IGraphicsFrame
+	{
+		int Width { get; }
+		int Height { get; }
+		ICharacterBuffer CharacterBuffer { get; }
+		IColorBuffer ColorBuffer { get; }
+	}
+}
