+++ GraphicsEngine/Scenes/MeshTestScene.cs
@@ -4,15 +4,15 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Windows.Input;
+using GraphicsEngine.Engine;
 using GraphicsEngine.Graphics;
-using GraphicsEngine.Graphics.Console;
 using GraphicsEngine.Math;
 using GraphicsEngine.Wavefront.Loaders;
 using GraphicsEngine.Win32;
 
 #endregion
 
-namespace GraphicsEngine.Engine
+namespace GraphicsEngine.Scenes
 {
 	public class MeshTestScene
 		: IScene
@@ -21,7 +21,7 @@ namespace GraphicsEngine.Engine
 		private IEnumerable<IMesh> meshes;
 		private float scaleFactor = 0.5f;
 		private readonly InputStateService inputStateService;
-		private readonly LazyRasterizer rasterizer;
+		private readonly IRasterizer rasterizer;
 		private readonly Transformation transformation;
 		private readonly string[] wavefrontObjectFilePaths = new[] {"Models\\triangle.obj", "Models\\cube.obj", "Models\\sphere.obj", "Models\\conf.obj", "Models\\gourd.obj", "Models\\link.obj", "Models\\monkey.obj", "Models\\bunny.obj", "Models\\f1.obj", "Models\\woman1.obj" };
 
@@ -30,7 +30,7 @@ namespace GraphicsEngine.Engine
 			Width = width;
 			Height = height;
 
-			rasterizer = new LazyRasterizer(Width, Height);
+			rasterizer = new Rasterizer(Width, Height);
 			inputStateService = new InputStateService();
 			transformation = new Transformation();
 
@@ -48,18 +48,18 @@ namespace GraphicsEngine.Engine
 			UpdateTransformation();
 		}
 
-		public void Draw()
+		public GraphicsFrame Rasterize()
 		{
-			rasterizer.ClearImage((byte) ' ', (byte) Kernel32Console.Colors.BACKGROUND_BLACK | Kernel32Console.Colors.FOREGROUND_GREY | Kernel32Console.Colors.FOREGROUND_INTENSITY);
+			rasterizer.ClearImage((byte)' ', (byte)Kernel32Console.Colors.BACKGROUND_BLACK | Kernel32Console.Colors.FOREGROUND_GREY | Kernel32Console.Colors.FOREGROUND_INTENSITY);
 
-			//rasterizer.DrawMeshFilled(transformation, meshes);//, Kernel32Console.Colors.FOREGROUND_RED, Rasterizer.ShadePixelChar1);
-			rasterizer.DrawMeshVertices(transformation, meshes, Kernel32Console.Colors.FOREGROUND_CYAN, (byte)'X');
-			//rasterizer.DrawMeshWired(transformation, meshes, Kernel32Console.Colors.FOREGROUND_GREEN);
+			//rasterizer.DrawMeshFilled(transformation, meshes);//, Kernel32Console.Colors.FOREGROUND_CYAN, Rasterizer.ShadePixelChar1);
+			//rasterizer.DrawMeshVertices(transformation, meshes, Kernel32Console.Colors.FOREGROUND_CYAN, (byte)'X');
+			rasterizer.DrawMeshWired(transformation, meshes, Kernel32Console.Colors.FOREGROUND_CYAN);
 
-			//rasterizer.DrawMeshCenters(transformation, meshes, Kernel32Console.Colors.FOREGROUND_GREEN);
+			rasterizer.DrawMeshCenters(transformation, meshes, Kernel32Console.Colors.FOREGROUND_RED);
 			//rasterizer.DrawMeshBoundingBox(transformation, meshes, Kernel32Console.Colors.FOREGROUND_MAGENTA);
 
-			//rasterizer.DrawAxes(Transformation.None, Kernel32Console.Colors.FOREGROUND_YELLOW);
+			rasterizer.DrawAxes(Transformation.None, Kernel32Console.Colors.FOREGROUND_YELLOW);
 
 			rasterizer.DrawStringHorizontal(Transformation.None, new Vector2(-Width / 2 + 1, Height / 2 - 2), string.Format("MODEL: '{0}'", currentWavefrontObjectFilePath));
 			rasterizer.DrawStringHorizontal(Transformation.None, new Vector2(-Width / 2 + 1, Height / 2 - 3), string.Format("# POLYGONS: {0}", meshes.Sum(mesh => mesh.Faces.Count())));
@@ -69,10 +69,7 @@ namespace GraphicsEngine.Engine
 			rasterizer.DrawStringHorizontal(Transformation.None, new Vector2(-Width / 2 + 1, 8), string.Format("ROTATION: {0}", transformation.Rotation));
 
 			rasterizer.DrawStringHorizontal(Transformation.None, new Vector2(-Width / 2 + 1, 6), string.Format("MODEL CENTER: {0}", transformation.Transform(meshes.First().Centers.Value)));
-		}
 
-		public ConsoleGraphicsFrame Rasterize()
-		{
 			var rasterizedFrame = rasterizer.Rasterize();
 			return rasterizedFrame;
 		}
@@ -87,6 +84,7 @@ namespace GraphicsEngine.Engine
 			wavefrontObjFileStream.Close();
 			var wavefrontObjToMeshConverter = new WavefrontObjToMeshConverter();
 			meshes = wavefrontObjToMeshConverter.ConvertToMesh(wavefrontObj);
+			meshes = MeshCenterer.Instance.Transform(meshes);
 			currentWavefrontObjectFilePath = wavefrontObjFilePath;
 			transformation.Reset();
 		}
