+++ GraphicsEngine/Scenes/RandomCharactersScene.cs
@@ -1,29 +1,28 @@
 #region Imports
 
 using System;
-using System.Threading;
+using GraphicsEngine.Engine;
 using GraphicsEngine.Graphics;
-using GraphicsEngine.Graphics.Console;
 using GraphicsEngine.Math;
 using GraphicsEngine.Win32;
 
 #endregion
 
-namespace GraphicsEngine.Engine
+namespace GraphicsEngine.Scenes
 {
 	public class RandomCharactersScene
 		: IScene
 	{
 		private static readonly short[] colors = new short[] {Kernel32Console.Colors.FOREGROUND_BLUE, Kernel32Console.Colors.FOREGROUND_CYAN, Kernel32Console.Colors.FOREGROUND_GREEN, Kernel32Console.Colors.FOREGROUND_MAGENTA, Kernel32Console.Colors.FOREGROUND_RED, Kernel32Console.Colors.FOREGROUND_YELLOW, Kernel32Console.Colors.FOREGROUND_GREY};
 		private readonly Random random = new Random(1000);
-		private readonly LazyRasterizer rasterizer;
+		private readonly IRasterizer rasterizer;
 
 		public RandomCharactersScene(int width, int height)
 		{
 			Width = width;
 			Height = height;
 
-			rasterizer = new LazyRasterizer(Width, Height);
+			rasterizer = new Rasterizer(Width, Height);
 		}
 
 		public int Width { get; }
@@ -33,7 +32,7 @@ namespace GraphicsEngine.Engine
 		{
 		}
 
-		public void Draw()
+		public GraphicsFrame Rasterize()
 		{
 			//Thread.Sleep(50);
 
@@ -41,16 +40,13 @@ namespace GraphicsEngine.Engine
 
 			for (int i = 0; i < 50; i++)
 			{
-				var randomX = random.Next(-Width/2, Width/2);
-				var randomY = random.Next(-Height/2, Height/2);
+				var randomX = random.Next(-Width / 2, Width / 2);
+				var randomY = random.Next(-Height / 2, Height / 2);
 				var randomColor = colors[random.Next(0, colors.Length - 1)];
 				var randomCharacter = (char)random.Next(char.MinValue, char.MaxValue);
 				rasterizer.DrawPoint(Transformation.None, new Vector2(randomX, randomY), randomColor, (byte)randomCharacter);
 			}
-		}
 
-		public ConsoleGraphicsFrame Rasterize()
-		{
 			var rasterizedFrame = rasterizer.Rasterize();
 			return rasterizedFrame;
 		}
