+++ GraphicsEngine/Scenes/RandomLinesScene.cs
@@ -2,28 +2,28 @@
 
 using System;
 using System.Threading;
+using GraphicsEngine.Engine;
 using GraphicsEngine.Graphics;
-using GraphicsEngine.Graphics.Console;
 using GraphicsEngine.Math;
 using GraphicsEngine.Win32;
 
 #endregion
 
-namespace GraphicsEngine.Engine
+namespace GraphicsEngine.Scenes
 {
 	public class RandomLinesScene
 		: IScene
 	{
 		private static readonly short[] colors = new short[] {Kernel32Console.Colors.FOREGROUND_BLUE, Kernel32Console.Colors.FOREGROUND_CYAN, Kernel32Console.Colors.FOREGROUND_GREEN, Kernel32Console.Colors.FOREGROUND_MAGENTA, Kernel32Console.Colors.FOREGROUND_RED, Kernel32Console.Colors.FOREGROUND_YELLOW, Kernel32Console.Colors.FOREGROUND_GREY};
 		private readonly Random random = new Random(1000);
-		private readonly LazyRasterizer rasterizer;
+		private readonly IRasterizer rasterizer;
 
 		public RandomLinesScene(int width, int height)
 		{
 			Width = width;
 			Height = height;
 
-			rasterizer = new LazyRasterizer(Width, Height);
+			rasterizer = new Rasterizer(Width, Height);
 		}
 
 		public int Width { get; }
@@ -33,11 +33,11 @@ namespace GraphicsEngine.Engine
 		{
 		}
 
-		public void Draw()
+		public GraphicsFrame Rasterize()
 		{
 			Thread.Sleep(50);
 
-			rasterizer.ClearImage((byte) ' ', (byte) Kernel32Console.Colors.BACKGROUND_BLACK | Kernel32Console.Colors.FOREGROUND_GREY | Kernel32Console.Colors.FOREGROUND_INTENSITY);
+			rasterizer.ClearImage((byte)' ', (byte)Kernel32Console.Colors.BACKGROUND_BLACK | Kernel32Console.Colors.FOREGROUND_GREY | Kernel32Console.Colors.FOREGROUND_INTENSITY);
 
 			for (var i = 0; i < 10; i++)
 			{
@@ -46,12 +46,9 @@ namespace GraphicsEngine.Engine
 				var randomX2 = random.Next(-Width / 2, Width / 2);
 				var randomY2 = random.Next(-Height / 2, Height / 2);
 				var randomColorIndex = random.Next(0, colors.Length);
-				rasterizer.DrawLine(Transformation.None, new Vector2(randomX1, randomY1), new Vector2(randomX2, randomY2), (short) (colors[randomColorIndex]), LazyRasterizer.HalfPixelChar);
+				rasterizer.DrawLine(Transformation.None, new Vector2(randomX1, randomY1), new Vector2(randomX2, randomY2), (short)(colors[randomColorIndex]), Rasterizer.HalfPixelChar);
 			}
-		}
 
-		public ConsoleGraphicsFrame Rasterize()
-		{
 			var rasterizedFrame = rasterizer.Rasterize();
 			return rasterizedFrame;
 		}
